// types/types.d.ts
export type CustomTheme = {
    primary: string;
    primaryContent: string;
    secondary: string;
    secondaryContent: string;
    accent: string;
    accentContent: string;
    neutral: string;
    neutralContent: string;
    base100: string;
    base200: string;
    base300: string;
    baseContent: string;
    info: string;
    infoContent: string;
    success: string;
    successContent: string;
    warning: string;
    warningContent: string;
    error: string;
    errorContent: string;
    roundedBox: string;
    roundedBtn: string;
    roundedBadge: string;
    animationBtn: string;
    animationInput: string;
    btnFocusScale: string;
    borderBtn: string;
    tabBorder: string;
    tabRadius: string;
};